DEVELOPMENT OF PRODUCTION TECHNOLOGY FOR FIBROUS, HEAT-INSULATING, FILTERING MATERIALS AND TECHNICAL PRODUCTS
This article presents the results of research on the influence of external factors on the quality of processed materials from the “Qizilqum” basalt deposits. During the experiments, mechanical cleaning and assessment of the chemical properties of basalt rock samples obtained from three different deposits were carried out. The obtained results demonstrated the possibility of extending the service life and increasing the corrosion resistance of equipment by washing basalt rocks. The article offers suggestions for the use of basalts as thermal insulation materials and ensuring their environmental safety.